feat(ui): show variant stock indicator in parts table (#5959)
